Add logging for submissions that take a long time to grade.
Passing GRADING_TIMEOUT to the http request is deceptive in that the transaction does not timeout
after the given time.  The grading timeout as it is currently used in the http request will time
out if we don't begin to receive a response from the server in the given time but the response
could take a long time to finish.