Swap around make and tox
This allows make to run you just the current virtualenv again while
allowing things like travis to run tox across all the virtualenvs.

This also means we don't need the long test command in two places that
might drift later.

Default MySQL and Rabbit to other devstack containers

The rest of this needs renaming and cleanup because we don't use JENKINS
but we do want to allow folks to have a test settings file with things
like matlab tokens.

Switch to a variant of travis where we bring up a docker cluster

I suspect this still needs help to make tox run all the versions.

We need to allow through DB/Rabbit for testing