Break apart code into separate worker classes
This lets us put the rabbit code in a subclass that's only
loaded when WABBITS is set to RabbitMQ.  Otherwise it uses
only a database.  Keep Worker() around as a shim to choose and later it
will return to being the top class.

Refactor get_single_unretired_submission since one needs
rabbit + database, the other needs just a simple database check.