Run tests twice to check database vs rabbit
This started off using ddt and override_settings to run a number of the
tests twice, but it was needed multiple times in many test files, and it
caused odd setup/teardown errors.  This was simpler except for the
mostly horrid temporary hack for coverage.  pytest's --cov-append didn't
merge .coveragerc, it just wrote new versions of the test in.  coverage
can union the two files.