It's very slow to query on these 4 columns
Build compound queries, previously we only had indexes on retired,
lms_callback_url and queue_name so queries would need to search all
submissions for a given queue or retired submissions, both of which
could be huge (and take minutes to query).  This should improve that.

There's also an update that runs across 2 columns and tends to lock a
lot of the table while doing so.