A submissions-per-queue management command
Useful to have xqueue tell you in a tabular printout if you have a bunch
of pending submissions.  Also able to send data to New Relic Insights.

In order to send metrics to New Relic, the management command needs the
app name and to ensure the license key is in the environment prior to
importing the agent.