Update coverage for our tests.
Move ext_interface tests out of lms_interface

Start to put in tests for the rest of the api endpoints.

Test 2 larger chunks of lms_interface

A annoying test to upload sufficient files to cause XQueue to use the
block of code that runs with it can't store file download URLs in the
database.
Since we use filesystem "urls" for this, they're always short and we
need to jump through hoops to approximate what happens when you upload a
few files to S3.

Note the struggle of testing external interface as written

We already test many submissions using method calls, but make one call
that tests the request handling code.

Add a test about queuelen and get_submission
This checks that when you pull a submission, it disappears from
the the queue count.

Two bugfixes:

Better handle totally invalid replies (rather than 500ing if the header
doesn't have the right keys)

Tests that run the consumer startup command

This noticed a bug from when I refactored all of this.
Workers no longer have ids, and I'd rather log the queue whose
consumer died and was restarted / exited than log PIDs.