This command was for putting things back in Rabbit
I misread some of the code earlier, but while trying to refactor some of
the tests, it seemed clear to me now that this is really just for taking
failed requests and putting them back on Rabbit.  Since failed
submissions will still be available for graders up until
MAX_NUMBER_OF_FAILURES before we auto-retire them, I don't see as much
used for this command.