Update Travis MySQL Image to 5.7 (#783)
Updated the devstack and Travis database settings to use MySQL 5.7 instead of 5.6. This triggered problems in one test due to how multiprocessing interacts with Django database connections, as described in https://stackoverflow.com/questions/24406092/when-using-multiprocessing-to-access-mysql-it-always-throws-the-following-error . The test was updated to close the parent process's connection before forking the sub-processes.

Also, nothing ever shut down the worker processes explicitly once they were started, so added that to the test grader stub.

Finally, coverage combine started failing with a "No data to combine" message, even though it doesn't seem like anything relevant has changed in this PR and it was previously working on master. The command is redundant anyway since pytest-cov does the combining into a single .coverage file before pytest exits, so it was removed.