build: Simplify xqueue test setup.
Previously we tested inside an outdated xqueue container but that
container no longer gets updates and so that strategy is out of date.
We update the tests to just test locally using github's capabilities to
manage the python version.

We could also use it to manage mysql and that's a future improvement we
can make but for now this should unblock the tests working again.  I'm
aiming to do the minimal necessary to fix the tests since this repo is
deprecated and will go away within the next year.